[OC-3564] update spec to not depend on ActiveSupport
Spec in question was improperly assuming the `blank?` method exists for 
Object instances.  This method is not part of core Ruby and actually 
only ships as part of ActiveSupport. ActiveSupport was being loaded when
by other specs that rely on treetop (which has ActiveSupport as a 
dependency). An equivalent test would be checking that a value is not 
nil? or empty?.